requirements.txt
absl-py==1.0.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
flatbuffers==2.0
fonttools==4.30.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
hydra-core==1.1.1
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.2.9
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
